Move Dagger statistics into a separate subpackage.
RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=260643012
